pyproject.toml
[project]
name = "deadlock-data-api"
version = "0.1.0"
description = ""
authors = [
    { name = "Manuel Raimann", email = "raimannma@outlook.de" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "cachetools>=5.5.0",
    "clickhouse-pool>=0.6.0",
    "fastapi>=0.115.4",
    "more-itertools>=10.6.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "protobuf>=5.28.3",
    "psycopg2-binary>=2.9.10",
    "redis[hiredis]>=5.2.0",
    "requests>=2.32.3",
    "retry>=0.9.2",
    "sentry-sdk[fastapi]>=2.18.0",
    "uuid>=1.30",
    "uvicorn>=0.34.0",
    "valveprotos-py",
]

[dependency-groups]
dev = [
    "ruff>=0.7.3",
]

[tool.uv.sources]
valveprotos-py = { git = "https://github.com/deadlock-api/valveprotos-py" }

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = [
    "I", # isort-like rules
    "UP", # pyupgrade-like rules
]
